+++ empty.h
@@ -13,7 +13,7 @@ class empty : public figure{
 public:
     
     empty(figure::color cc);
-    bool can_move(int x1, int x2, int y1, int y2);
+    bool can_move(short int x1, short int x2, short int y1, short int y2);
     virtual ~empty();
 private:
