+++ pawn.h
@@ -15,8 +15,8 @@ public:
     pawn(figure::color cc);
 
     virtual ~pawn();
-    bool can_move(int x1, int x2, int y1, int y2);
-    bool can_capture(int x1, int x2, int y1, int y2);
+    bool can_move(short int x1, short int x2, short int y1, short int y2);
+    bool can_capture(short int x1, short int x2, short int y1, short int y2);
 private:
     
 };
