+++ queen.h
@@ -13,7 +13,7 @@ class queen : public figure{
 public:
    
     queen(figure::color cc);
-    virtual bool can_move(int x1, int x2, int y1, int y2);
+    virtual bool can_move(short int x1, short int x2, short int y1, short int y2);
     virtual ~queen();
 private:
