+++ rook.cpp
@@ -15,7 +15,7 @@ rook::rook(figure::color cc):figure(cc){sign ='W';}
 rook::~rook() {
 }
 
-bool rook::can_move(int x1, int x2, int y1, int y2) {
+bool rook::can_move(short int x1, short int x2, short int y1, short int y2) {
  
         if ( 
              (
